Issue 31: decode error (#158)
This is a very, perhaps overly, careful implementation of decoding
contract errors. It is implemented only for one case, namely
Contract::Revert errors, and only if the format is as expected
(according to the ABI). We can roll out this solution to other error
types, but I need to check the ABI specs.